refactor(capabilities): split store into a submodule
Closes #2162
